+++ src/platform/window_sdl.rs
@@ -92,6 +92,8 @@ impl Window {
             window_builder.input_grabbed();
         }
 
+        sdl.mouse()
+            .set_relative_mouse_mode(settings.relative_mouse_mode);
         sdl.mouse().show_cursor(settings.show_mouse);
 
         let mut sdl_window = window_builder
@@ -132,8 +134,9 @@ impl Window {
             .gl_create_context()
             .map_err(TetraError::PlatformError)?;
 
-        let gl_ctx =
-            GlowContext::from_loader_function(|s| video_sys.gl_get_proc_address(s) as *const _);
+        let gl_ctx = unsafe {
+            GlowContext::from_loader_function(|s| video_sys.gl_get_proc_address(s) as *const _)
+        };
 
         video_sys
             .gl_set_swap_interval(if settings.vsync {
@@ -288,6 +291,16 @@ impl Window {
         self.sdl_window.grab()
     }
 
+    pub fn set_relative_mouse_mode(&mut self, relative_mouse_mode: bool) {
+        self.sdl
+            .mouse()
+            .set_relative_mouse_mode(relative_mouse_mode);
+    }
+
+    pub fn is_relative_mouse_mode(&self) -> bool {
+        self.sdl.mouse().relative_mouse_mode()
+    }
+
     pub fn get_clipboard_text(&self) -> Result<String> {
         self.video_sys
             .clipboard()
@@ -409,10 +422,14 @@ where
                 }
             }
 
-            SdlEvent::MouseMotion { x, y, .. } => {
+            SdlEvent::MouseMotion {
+                x, y, xrel, yrel, ..
+            } => {
                 let position = Vec2::new(x as f32, y as f32);
+                let delta = Vec2::new(xrel as f32, yrel as f32);
+
                 input::set_mouse_position(ctx, position);
-                state.event(ctx, Event::MouseMoved { position })?;
+                state.event(ctx, Event::MouseMoved { position, delta })?;
             }
 
             SdlEvent::MouseWheel {
@@ -582,8 +599,6 @@ where
 
     Ok(())
 }
-
-// TODO: Replace these with TryFrom once we're on a high enough minimum Rust version?
 
 fn into_mouse_button(button: SdlMouseButton) -> Option<MouseButton> {
     match button {